Operator#fetchRedundancyFactor
